pyproject.toml
[tool.poetry]
name = "dsop-api-spesifikasjoner"
version = "0.1.0"
description = "A collection of openAPI specifications for the DSOP-banks."
authors = ["Stig B. Dørmænen <stigbd@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^5.3.1"
click = "^7.1.2"
jsonpickle = "^1.4.1"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
black = "^20.8b1"
flake8 = "^3.8.4"
flake8-black = "^0.2.1"
mypy = "^0.782"
pytype = {version = "^2020.2.20", python = "3.7"}
flake8-annotations = "^2.4.1"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^20.1.4"
flake8-docstrings = "^1.5.0"
flake8-import-order = "^0.18.1"
darglint = "^1.5.5"
flake8-assertive = "^1.2.1"
pytest-cov = "^2.10.1"
deepdiff = "^5.0.2"
pytest-mock = "^3.3.1"

[tool.poetry.scripts]
dsop_api_spesifikasjoner = "dsop_api_spesifikasjoner.generateSpecification:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = false
source = ["dsop_api_spesifikasjoner"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
